requirements.txt
numpy==1.24.3
Flask==2.2.2
Werkzeug==2.2.2
cohere==4.39.0
pandas
umap-learn
altair
scikit-learn
annoy
tqdm
datasets
python-dotenv
pinecone-client
datasets